+++ arm9/source/qsort.h
@@ -1,6 +1,6 @@
 #ifndef __QSORT_H__
 #define __QSORT_H__
 
-void qsort(void *aa, size_t n, size_t es, int (*cmp)(const void *, const void *));
+void qsort(void *aa, int n, int es, int (*cmp)(const void *, const void *));
 
 #endif
